# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"AuthEntity": {
				"properties": {
					"refreshToken": {
						"type": "string"
					},
					"authorizationToken": {
						"type": "string"
					}
				},
				"required": [
					"refreshToken",
					"authorizationToken"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SignInType": {
				"properties": {
					"pass": {
						"type": "string"
					},
					"user": {
						"type": "string"
					}
				},
				"required": [
					"pass",
					"user"
				],
				"type": "object"
			},
			"UserEntity": {
				"properties": {
					"id": {
						"type": "string"
					},
					"_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"nickname": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"team": {
						"type": "string"
					},
					"lastConnection": {
						"type": "string",
						"format": "date-time"
					},
					"token": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"lastName",
					"nickname",
					"email",
					"team"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "pokeapi-rocketfy-server",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/auth/signin": {
			"post": {
				"operationId": "SignIn",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthEntity"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SignInType"
							}
						}
					}
				}
			}
		},
		"/auth/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserEntity"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserEntity"
							}
						}
					}
				}
			}
		},
		"/auth/user": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserEntity"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/auth/refresh": {
			"get": {
				"operationId": "Refresh",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthEntity"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}